requirements.txt
pygame-ce>=2.1.2
